requirements.txt
fastapi
uvicorn


pyjwt
python-decouple